+++ libp2p/peerid.nim
@@ -215,11 +215,11 @@ proc getValue*(data: var ProtoBuffer, field: int, value: var PeerID): int {.
     else:
       value = pid
 
-proc write*(pb: var ProtoBuffer, field: int64, pid: PeerID) =
+proc write*(pb: var ProtoBuffer, field: int, pid: PeerID) =
   ## Write PeerID value ``peerid`` to object ``pb`` using ProtoBuf's encoding.
   write(pb, field, pid.data)
 
-proc getField*(pb: ProtoBuffer, field: int64, pid: var PeerID): bool =
+proc getField*(pb: ProtoBuffer, field: int, pid: var PeerID): bool =
   ## Read ``PeerID`` from ProtoBuf's message and validate it
   var buffer: seq[byte]
   var peerId: PeerID
